moved content in page folder back to components folder
