+++ client/src/components/Login/BusinessLogin.tsx
@@ -68,7 +68,7 @@ export default function BusinessLogin ({ setAuth, setUserType }: AuthProps): JSX
         </Button>
         <Grid container>
           <Grid item>
-            <Link href="/signup" variant="body2">
+            <Link href="/client/src/components/Signup" variant="body2">
               Don't have an account? Sign Up
             </Link>
           </Grid>
